refactor(App.js): move all functions inside Game class as methods
BREAKING CHANGE: Game class in Game.js file contains all the methods to run the lottery game
